the entire loop in lib.py
